backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
sqlmodel==0.0.22
psycopg-pool==3.2.3
psycopg[binary]==3.2.3
pgvector==0.3.2
google-genai==1.0.0
PyMuPDF==1.24.10
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.3
